Kernel gets pid=0 + can have threads
Major change to process and scheduler architecture. When the kernel
starts up and process is initialized, it gets accounted for as PID=0

We support processes sharing memory space and also not having any memory
space, which enables effectively same-group threading as well as kernel
threads.

Processes that belong to the same memory space will generally have the
same PGID.

Added a wait queue as the beginnings of synchronization primitives, and
made exit status waiting use that. (Keyboard might also benefit in the
future)

Also a handy macro that makes it easier to join a wait queue and check a
condition on wakeup - see `wait!`